vkd3d: Introduce interface for creating d3d12 resources from VkImages
Motivation for adding this is to support OpenXR extension
XR_KHR_D3D12_enable.

The way OpenXR is designed, the application gets an array of swapchain
images from the runtime, and renders into them. (Unlike OpenVR, with
which the application will submit textures to be displayed in the VR
environment). This means we need to convert the VkImages we get on the
unix side to ID3D12Resources on the PE side.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>